refactor(rpc): move add transaction handlers to the v0.9 module
